+++ src/app/recipe/page.tsx
@@ -1,15 +1,24 @@
 import Header from '@/components/layout/Header';
+import RecipeImage from '@/components/recipe/RecipeImage';
+import RecipeTitle from '@/components/recipe/RecipeTitle';
+import IngredientsList from '@/components/recipe/IngredientsList';
+import StepList from '@/components/recipe/StepList';
+import BottomNavigation from '@/components/recipe/BottomNavigation';
 
-export default function HomePage() {
+export default function Recipe() {
   return (
-    <div className="p-4">
+    <div className="min-h-screen bg-white">
       <Header />
 
-      <main className="space-y-6">
-        <section>
-          <h2 className="text-xl font-bold mb-4">home</h2>
-        </section>
+      {/* Main Content */}
+      <main className="px-4 py-4 pb-20">
+        <RecipeImage />
+        <RecipeTitle />
+        <IngredientsList />
+        <StepList />
       </main>
+
+      <BottomNavigation />
     </div>
   );
 }
