+++ src/app/search/page.tsx
@@ -1,0 +1,15 @@
+import Header from '@/components/layout/Header';
+
+export default function Search() {
+  return (
+    <div className="p-4">
+      <Header />
+
+      <main className="space-y-6">
+        <section>
+          <h2 className="text-xl font-bold mb-4">search</h2>
+        </section>
+      </main>
+    </div>
+  );
+}
